+++ buildLogic/multiplatform-convention/src/main/kotlin/app/cash/sqldelight/multiplatform/MultiplatformConventions.kt
@@ -27,13 +27,6 @@ class MultiplatformConventions : Plugin<Project> {
         }
 
         if (it is KotlinJsTarget) {
-          it.nodejs {
-            testTask {
-              useMocha {
-                timeout = "5s"
-              }
-            }
-          }
           it.browser {
             testTask {
               useKarma {
@@ -58,11 +51,9 @@ class MultiplatformConventions : Plugin<Project> {
       }
 
       iosX64()
-      iosArm32()
       iosArm64()
       tvosX64()
       tvosArm64()
-      watchosX86()
       watchosX64()
       watchosArm32()
       watchosArm64()