+++ drivers/driver-test/src/commonMain/kotlin/com/squareup/sqldelight/driver/test/QueryTest.kt
@@ -174,7 +174,7 @@ abstract class QueryTest {
 
   private fun testDataQuery(): Query<TestData> {
     return object : Query<TestData>(mapper) {
-      override fun <R> execute(mapper: (SqlCursor) -> R): QueryResult<R> {
+      override fun <R> execute(mapper: (SqlCursor) -> QueryResult<R>): QueryResult<R> {
         return driver.executeQuery(0, "SELECT * FROM test", mapper, 0, null)
       }