+++ extensions/rxjava2-extensions/src/test/kotlin/app/cash/sqldelight/rx2/RecordingObserver.kt
@@ -16,6 +16,7 @@
 package app.cash.sqldelight.rx2
 
 import app.cash.sqldelight.Query
+import app.cash.sqldelight.db.QueryResult
 import com.google.common.truth.Truth.assertThat
 import com.google.common.truth.Truth.assertWithMessage
 import io.reactivex.observers.DisposableObserver
@@ -37,9 +38,9 @@ internal class RecordingObserver(val numberOfColumns: Int) : DisposableObserver<
   override fun onNext(value: Query<*>) {
     val allRows = value.execute { cursor ->
       val data = mutableListOf<List<String?>>()
-      while (cursor.next())
+      while (cursor.next().value)
         data.add((0 until numberOfColumns).map(cursor::getString))
-      data
+      QueryResult.Value(data)
     }.value
     events.add(allRows)
   }