+++ sqldelight-compiler/integration-tests/src/test/kotlin/com/example/Myftstable2.kt
@@ -1,0 +1,3 @@
+package com.example
+
+data class Myftstable2(val something: String?, val nice: String?)