+++ sqldelight-compiler/src/main/kotlin/app/cash/sqldelight/core/compiler/SelectQueryGenerator.kt
@@ -229,8 +229,9 @@ class SelectQueryGenerator(
       val tablesObserved = query.tablesObserved
       if (tablesObserved.isNullOrEmpty()) {
         function.addCode(
-          "return %T(${query.id}, $DRIVER_NAME, %S, %S, %S)%L",
+          "return %T(%L, $DRIVER_NAME, %S, %S, %S)%L",
           QUERY_TYPE,
+          query.id,
           query.statement.containingFile.name,
           query.name,
           query.statement.rawSqlText(),
@@ -238,8 +239,9 @@ class SelectQueryGenerator(
         )
       } else {
         function.addCode(
-          "return %T(${query.id}, %L, $DRIVER_NAME, %S, %S, %S)%L",
+          "return %T(%L, %L, $DRIVER_NAME, %S, %S, %S)%L",
           QUERY_TYPE,
+          query.id,
           queryKeys(tablesObserved),
           query.statement.containingFile.name,
           query.name,
@@ -305,9 +307,11 @@ class SelectQueryGenerator(
 
     val genericResultType = TypeVariableName("R")
     val createStatementFunction = FunSpec.builder(EXECUTE_METHOD)
-      .addModifiers(OVERRIDE)
+      .addModifiers(
+        OVERRIDE,
+      )
       .addTypeVariable(genericResultType)
-      .addParameter(MAPPER_NAME, LambdaTypeName.get(parameters = arrayOf(CURSOR_TYPE), returnType = genericResultType))
+      .addParameter(MAPPER_NAME, LambdaTypeName.get(parameters = arrayOf(CURSOR_TYPE), returnType = QUERY_RESULT_TYPE.parameterizedBy(genericResultType)))
       .returns(QUERY_RESULT_TYPE.parameterizedBy(genericResultType))
       .addCode(executeBlock())