+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/queries/InterfaceGeneration.kt
@@ -7,6 +7,7 @@ import app.cash.sqldelight.core.compiler.TableInterfaceGenerator
 import app.cash.sqldelight.dialects.postgresql.PostgreSqlDialect
 import app.cash.sqldelight.test.util.FixtureCompiler
 import app.cash.sqldelight.test.util.withInvariantLineSeparators
+import app.cash.sqldelight.test.util.withUnderscores
 import com.google.common.truth.Truth.assertThat
 import com.google.common.truth.Truth.assertWithMessage
 import org.junit.Rule
@@ -829,7 +830,7 @@ class InterfaceGeneration {
       |      driver.removeListener(listener, arrayOf("song"))
       |    }
       |
-      |    public override fun <R> execute(mapper: (SqlCursor) -> R): QueryResult<R> =
+      |    public override fun <R> execute(mapper: (SqlCursor) -> QueryResult<R>): QueryResult<R> =
       |        driver.executeQuery(null,
       |        ""${'"'}SELECT * FROM song WHERE album_id ${'$'}{ if (album_id == null) "IS" else "=" } ?""${'"'}, mapper,
       |        1) {
@@ -904,14 +905,14 @@ class InterfaceGeneration {
       |  }
       |
       |  public fun insertSubscription(user_id2: Int): Unit {
-      |    driver.execute(${result.compiledFile.namedMutators[0].id}, ""${'"'}
+      |    driver.execute(${result.compiledFile.namedMutators[0].id.withUnderscores}, ""${'"'}
       |        |INSERT INTO subscriptionEntity(user_id2)
       |        |VALUES (?)
       |        ""${'"'}.trimMargin(), 1) {
       |          check(this is JdbcPreparedStatement)
       |          bindLong(0, user_id2.toLong())
       |        }
-      |    notifyQueries(${result.compiledFile.namedMutators[0].id}) { emit ->
+      |    notifyQueries(${result.compiledFile.namedMutators[0].id.withUnderscores}) { emit ->
       |      emit("subscriptionEntity")
       |    }
       |  }
@@ -928,8 +929,8 @@ class InterfaceGeneration {
       |      driver.removeListener(listener, arrayOf("userEntity"))
       |    }
       |
-      |    public override fun <R> execute(mapper: (SqlCursor) -> R): QueryResult<R> =
-      |        driver.executeQuery(${result.compiledFile.namedQueries[0].id}, ""${'"'}
+      |    public override fun <R> execute(mapper: (SqlCursor) -> QueryResult<R>): QueryResult<R> =
+      |        driver.executeQuery(${result.compiledFile.namedQueries[0].id.withUnderscores}, ""${'"'}
       |    |WITH inserted_ids AS (
       |    |  INSERT INTO userEntity(slack_user_id)
       |    |  VALUES (?)
@@ -1042,8 +1043,8 @@ class InterfaceGeneration {
       |      driver.removeListener(listener, arrayOf("item"))
       |    }
       |
-      |    public override fun <R> execute(mapper: (SqlCursor) -> R): QueryResult<R> =
-      |        driver.executeQuery(${query.id}, ""${'"'}
+      |    public override fun <R> execute(mapper: (SqlCursor) -> QueryResult<R>): QueryResult<R> =
+      |        driver.executeQuery(${query.id.withUnderscores}, ""${'"'}
       |    |WITH RECURSIVE
       |    |descendants AS (
       |    |    SELECT id, parent_id