+++ sqldelight-gradle-plugin/src/test/kotlin/app/cash/sqldelight/WithCommonConfiguration.kt
@@ -11,8 +11,5 @@ internal fun GradleRunner.withCommonConfiguration(projectRoot: File): GradleRunn
       |
     """.trimMargin(),
   )
-  File(projectRoot, "local.properties").apply {
-    if (!exists()) writeText("sdk.dir=${androidHome()}\n")
-  }
   return withProjectDir(projectRoot)
 }