+++ sqldelight-idea-plugin/testData/unused-query-inspection/Example.kt
@@ -1,0 +1,9 @@
+package com.example
+
+import kotlin.Long
+import kotlin.String
+
+public data class Example(
+  public val id: Long,
+  public val name: String,
+)