+++ sqldelight-idea-plugin/testData/unused-query-inspection/Main.kt
@@ -1,0 +1,6 @@
+package com.example
+
+fun main() {
+  val someTableQueries = SomeTableQueries()
+  someTableQueries.selectAll()
+}