+++ sqldelight-idea-plugin/testData/unused-query-inspection/Query.kt
@@ -1,0 +1,5 @@
+class Query<T>(val value: T) {
+  fun executeAsOneOrNull(): T {
+    return value
+  }
+}