+++ CHANGELOG.md
@@ -26,6 +26,14 @@
 - [Compiler] Fix SelectQueryGenerator to support bind expression without parenthesis (by [Doogie Min][bellatoris])
 - [Compiler] Fix duplicate generation of ${name}Indexes variables when using transactions (by [Andreas Sacher][sachera])
 
+## [1.5.5] - 2023-01-20
+
+This is a compatibility release for Kotlin 1.8 and IntelliJ 2021+, supporting JDK 17.
+
+## [1.5.4] - 2022-10-06
+
+This is a compatibility update for Kotlin 1.7.20 and AGP 7.3.0.
+
 ## [2.0.0-alpha04] - 2022-10-03
 
 ### Breaking Changes